+++ CHANGELOG.rst
@@ -1,7 +1,8 @@
 1.3.22 (unreleased)
 -------------------
 
-- Nothing changed yet.
+- Add default head endpoint
+  [vangheem]
 
 
 1.3.21 (2017-09-27)